setting :: Configuration 중복값 수정
