--- Plugins/DependencyPlugin/Plugin.swift
@@ -1,3 +1,0 @@
-import ProjectDescription
-
-let plugin = Plugin(name: "DependencyPlugin")
