+++ Tuist/Plugins/ConfigurationPlugin/ProjectDescriptionHelpers/Configuration+Ext.swift
