+++ Tuist/Plugins/ConfigurationPlugin/ProjectDescriptionHelpers/Path+XCConfig.swift
