+++ Tuist/Plugins/DependencyPlugin/ProjectDescriptionHelpers/Dependency+SPM.swift
@@ -1,29 +1,31 @@
-
 import ProjectDescription
 
 public extension TargetDependency {
-    struct SPM {}
+    enum SPM {}
 }
 
 public extension TargetDependency.SPM {
-    static let Moya = TargetDependency.external(name: "Moya")
-    static let RxMoya = TargetDependency.external(name: "RxMoya")
     static let RxSwift = TargetDependency.external(name: "RxSwift")
     static let RxCocoa = TargetDependency.external(name: "RxCocoa")
     static let SnapKit = TargetDependency.external(name: "SnapKit")
     static let Then = TargetDependency.external(name: "Then")
-    static let Kingfisher = TargetDependency.external(name: "Kingfisher")
-    static let KeychainSwift = TargetDependency.external(name: "KeychainSwift")
     static let RxFlow = TargetDependency.external(name: "RxFlow")
-    static let RxGesture = TargetDependency.external(name: "RxGesture")
     static let Swinject = TargetDependency.external(name: "Swinject")
+    static let Moya = TargetDependency.external(name: "Moya")
+    static let RxMoya = TargetDependency.external(name: "RxMoya")
     static let Lottie = TargetDependency.external(name: "Lottie")
-    static let FSCalendar = TargetDependency.external(name: "FSCalendar")
-    static let RxDataSources = TargetDependency.external(name: "RxDataSources")
-    static let PiCKiOSDesignSystem = TargetDependency.external(name: "PiCK_iOS_DesignSystem")
-    static let FirebaseMessaging = TargetDependency.external(name: "FirebaseMessaging")
-    static let Starscream = TargetDependency.external(name: "Starscream")
+    static let kingfisher = TargetDependency.external(name: "Kingfisher")
+    static let KeychainSwift = TargetDependency.external(name: "KeychainSwift")
     static let ReactorKit = TargetDependency.external(name: "ReactorKit")
+    static let RxGesture = TargetDependency.external(name: "RxGesture")
     static let Pulse = TargetDependency.external(name: "Pulse")
+    static let PulseUI = TargetDependency.external(name: "PulseUI")
     static let ComposableArchitecture = TargetDependency.external(name: "ComposableArchitecture")
+//    static let DGCharts = TargetDependency.external(name: "DGCharts")
+    static let PDS = TargetDependency.external(name: "PiCK_iOS_DesignSystem")
+    static let FCM = TargetDependency.package(product: "FirebaseMessaging")
+}
+
+public extension Package {
+    static let FCM = Package.remote(url: "https://github.com/firebase/firebase-ios-sdk", requirement: .upToNextMajor(from: "10.0.0"))
 }
